requirements.txt
aiohttp>=3.8.1
alpaca-trade-api==3.2.0
authlib==1.2.1
bcrypt==4.1.2
fastapi>=0.115
fastapi-users[sqlalchemy]==15.0.1
Flask==2.3.2
gunicorn==20.1.0
httpx
requests>=2.32.0
itsdangerous==2.1.2
jinja2==3.1.2
langgraph>=0.1.3
matplotlib
msgpack>=1.0.4
azure-storage-blob>=12.23.0
azure-identity>=1.19.0
azure-communication-email>=1.0.0
numpy==1.26.4
openai>=1.50.0
pandas==2.2.2
pandas-ta @ git+https://github.com/0xAVX/pandas-ta.git
passlib[bcrypt]==1.7.4
pip>=25.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.20
scipy==1.13.0
setuptools>=80.0
sqlalchemy==2.0.23
stripe==8.10.0
uvicorn==0.23.2
wheel>=0.42
yfinance